+++ src/Http/Middleware/VerifyWebhookSignature.php
@@ -3,46 +3,20 @@
 namespace Laravel\Cashier\Http\Middleware;
 
 use Closure;
-use Illuminate\Contracts\Config\Repository as Config;
-use Illuminate\Contracts\Foundation\Application;
-use Stripe\Error\SignatureVerification;
+use Stripe\Exception\SignatureVerificationException;
 use Stripe\WebhookSignature;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class VerifyWebhookSignature
 {
-    /**
-     * The application instance.
-     *
-     * @var \Illuminate\Contracts\Foundation\Application
-     */
-    protected $app;
-
-    /**
-     * The configuration repository instance.
-     *
-     * @var \Illuminate\Contracts\Config\Repository
-     */
-    protected $config;
-
-    /**
-     * Create a new middleware instance.
-     *
-     * @param  \Illuminate\Contracts\Foundation\Application  $app
-     * @param  \Illuminate\Contracts\Config\Repository  $config
-     * @return void
-     */
-    public function __construct(Application $app, Config $config)
-    {
-        $this->app = $app;
-        $this->config = $config;
-    }
-
     /**
      * Handle the incoming request.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  \Closure  $next
      * @return \Illuminate\Http\Response
+     *
+     * @throws \Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException
      */
     public function handle($request, Closure $next)
     {
@@ -50,11 +24,11 @@ class VerifyWebhookSignature
             WebhookSignature::verifyHeader(
                 $request->getContent(),
                 $request->header('Stripe-Signature'),
-                $this->config->get('cashier.webhook.secret'),
-                $this->config->get('cashier.webhook.tolerance')
+                config('cashier.webhook.secret'),
+                config('cashier.webhook.tolerance')
             );
-        } catch (SignatureVerification $exception) {
-            $this->app->abort(403);
+        } catch (SignatureVerificationException $exception) {
+            throw new AccessDeniedHttpException($exception->getMessage(), $exception);
         }
 
         return $next($request);